+++ server/app/include/trantor/player.h
@@ -33,7 +33,7 @@ struct player_s {
     pcmd_executor_t pcmd_exec;
     player_t *incantator;
     unsigned int n;
-    double time_left;
+    float time_left;
     bool is_dead;
 };
